requirements.txt
# Backend API dependencies
fastapi
uvicorn[standard]
gunicorn
neo4j
faiss-cpu
sentence-transformers
numpy
python-dotenv
requests
pydantic
langdetect
html2text
ftfy
spacy
